Extract out test case editor
