+++ frontend/src/components/problem/editor/ProblemDisplay.tsx
@@ -47,6 +47,7 @@ import ProblemTags from '../ProblemTags';
 import { useAppSelector, useMousePosition, useProblemEditable } from '../../../util/Hook';
 import { ProblemHelpModal } from '../../core/HelpModal';
 import OptionsPanel from './OptionsPanel';
+import TestCaseEditor from './TestCaseEditor';
 
 const defaultDescription: string = [
   'Replace this line with a short description.',
@@ -71,40 +72,14 @@ const TopButtonsContainer = styled.div`
   margin-left: auto;
 `;
 
-const SettingsContainerRelative = styled(SettingsContainer)`
-  position: relative;
-`;
-
 const SettingsContainerHighPadding = styled(SettingsContainer)`
   padding: 3rem;
-`;
-
-const FlexBareContainerMarginBottom = styled(FlexBareContainer)`
-  margin-bottom: 10px;
 `;
 
 const TitleDescriptionSeparator = styled.hr`
   margin: 1rem 0;
   border: none;
   border-top: 1px solid ${({ theme }) => theme.colors.border};
-`;
-
-const InputContainer = styled.div`
-  margin-right: 10%;
-`;
-
-const HiddenContainer = styled.div`
-  margin-right: 5%;
-`;
-
-const ExplanationContainer = styled.div`
-  margin-right: 5%;
-  flex: 2;
-`;
-
-const MarginLeftRightAutoContainer = styled.div`
-  margin-right: 5%;
-  margin-left: auto;
 `;
 
 const StyledMarkdownEditor = styled(MarkdownEditor)`
@@ -116,10 +91,6 @@ const StyledMarkdownEditor = styled(MarkdownEditor)`
   .ProseMirror > p, blockquote, h1, h2, h3, ul, ol, table {
     color: ${({ theme }) => theme.colors.text};
   }
-`;
-
-const NoMarginTopText = styled(Text)`
-  margin-top: 0;
 `;
 
 const TextButtonLink = styled(TextButton)`
@@ -142,14 +113,6 @@ type ProblemDisplayParams = {
   editMode: boolean,
 };
 
-// a little function to help us with reordering the result
-const reorder = (list: TestCase[], startIndex: number, endIndex: number): TestCase[] => {
-  const result = Array.from(list);
-  const [removed] = result.splice(startIndex, 1);
-  result.splice(endIndex, 0, removed);
-  return result;
-};
-
 function ProblemDisplay(props: ProblemDisplayParams) {
   const {
     problem, onClick, actionText, editMode,
@@ -167,21 +130,6 @@ function ProblemDisplay(props: ProblemDisplayParams) {
   // Variable used to force refresh the editor.
   const [refreshEditor, setRefreshEditor] = useState<number>(0);
 
-  const onDragEnd = (result: any) => {
-    // dropped outside the list
-    if (!result.destination) {
-      return;
-    }
-
-    const newTestCases: TestCase[] = reorder(
-      newProblem.testCases,
-      result.source.index,
-      result.destination.index,
-    );
-
-    setNewProblem({ ...newProblem, testCases: newTestCases });
-  };
-
   // Handle updating of normal text fields
   const handleChange = (e: any) => {
     const { name, value } = e.target;
@@ -194,50 +142,13 @@ function ProblemDisplay(props: ProblemDisplayParams) {
   // Handle description change
   const handleDescriptionChange = (value: string) => handleChange({ target: { name: 'description', value } });
 
-  // Handle updating of test case
-  const handleTestCaseChange = (index: number, id: string, input: string,
-    output: string, hidden: boolean, explanation: string) => {
-    setNewProblem({
-      ...newProblem,
-      testCases: newProblem.testCases.map((testCase, i) => {
-        if (index === i) {
-          return {
-            id,
-            input,
-            output,
-            hidden,
-            explanation,
-          };
-        }
-        return testCase;
-      }),
-    });
-  };
-
-  // Handle adding a new test case for this problem
-  const addTestCase = () => {
-    setNewProblem({
-      ...newProblem,
-      testCases: [...newProblem.testCases, {
-        id: generateRandomId(), input: '0', output: '0', hidden: false, explanation: '',
-      }],
-    });
-  };
-
-  // Handle deleting a test case for this problem
-  const deleteTestCase = (index: number) => {
-    setNewProblem({
-      ...newProblem,
-      testCases: newProblem.testCases.filter((_, i) => index !== i),
-    });
-  };
-
   return (
     <>
       <ProblemHelpModal
         show={helpModal}
         exitModal={() => setHelpModal(false)}
       />
+
       <MainContent>
         <FlexBareContainer>
           <SmallHeaderText>Problem</SmallHeaderText>
@@ -297,114 +208,10 @@ function ProblemDisplay(props: ProblemDisplayParams) {
 
         {editMode && problemEditable
           ? (
-            <>
-              <SmallHeaderText>Test Cases</SmallHeaderText>
-              <DragDropContext onDragEnd={onDragEnd}>
-                <Droppable droppableId="droppable">
-                  {(providedDroppable: any) => (
-                    <div
-                      {...providedDroppable.droppableProps}
-                      ref={providedDroppable.innerRef}
-                    >
-                      {newProblem.testCases.map((testCase, index) => (
-                        <Draggable key={testCase.id} draggableId={testCase.id} index={index}>
-                          {(providedDraggable: any) => (
-                            <div
-                              ref={providedDraggable.innerRef}
-                              {...providedDraggable.draggableProps}
-                              {...providedDraggable.dragHandleProps}
-                              style={providedDraggable.draggableProps.style}
-                            >
-                              <SettingsContainerRelative>
-                                <LabelAbsoluteText>
-                                  #
-                                  {index + 1}
-                                </LabelAbsoluteText>
-                                <FlexBareContainer>
-                                  <InputContainer>
-                                    <NoMarginTopText>Input</NoMarginTopText>
-                                    <FixedTextArea
-                                      value={newProblem.testCases[index].input}
-                                      onChange={(e) => {
-                                        const current = newProblem.testCases[index];
-                                        handleTestCaseChange(index, current.id, e.target.value,
-                                          current.output, current.hidden, current.explanation);
-                                      }}
-                                    />
-                                  </InputContainer>
-                                  <MarginLeftRightAutoContainer>
-                                    <NoMarginTopText>Output</NoMarginTopText>
-                                    <FixedTextArea
-                                      value={newProblem.testCases[index].output}
-                                      onChange={(e) => {
-                                        const current = newProblem.testCases[index];
-                                        handleTestCaseChange(index, current.id,
-                                          current.input, e.target.value,
-                                          current.hidden, current.explanation);
-                                      }}
-                                    />
-                                  </MarginLeftRightAutoContainer>
-                                </FlexBareContainer>
-
-                                <FlexBareContainerMarginBottom>
-                                  <ExplanationContainer>
-                                    <Text>Explanation (optional)</Text>
-                                    <FixedTextArea
-                                      value={newProblem.testCases[index].explanation}
-                                      onChange={(e) => {
-                                        const current = newProblem.testCases[index];
-                                        handleTestCaseChange(index,
-                                          current.id, current.input,
-                                          current.output, current.hidden,
-                                          e.target.value);
-                                      }}
-                                    />
-                                  </ExplanationContainer>
-                                </FlexBareContainerMarginBottom>
-
-                                <FlexBareContainer>
-                                  <HiddenContainer>
-                                    <label htmlFor={`problem-hidden-${index}`}>
-                                      Hidden
-                                      <CheckboxInput
-                                        id={`problem-hidden-${index}`}
-                                        checked={newProblem.testCases[index].hidden}
-                                        onChange={(e) => {
-                                          const current = newProblem.testCases[index];
-                                          handleTestCaseChange(index,
-                                            current.id, current.input,
-                                            current.output, e.target.checked, current.explanation);
-                                        }}
-                                      />
-                                    </label>
-                                  </HiddenContainer>
-
-                                  <MarginLeftRightAutoContainer>
-                                    <RedTextButton
-                                      onClick={() => deleteTestCase(index)}
-                                    >
-                                      Delete
-                                    </RedTextButton>
-                                  </MarginLeftRightAutoContainer>
-                                </FlexBareContainer>
-                              </SettingsContainerRelative>
-                            </div>
-                          )}
-                        </Draggable>
-                      ))}
-                    </div>
-                  )}
-                </Droppable>
-              </DragDropContext>
-              {newProblem.testCases.length ? (
-                <Text>No test cases. At least 1 is required in order to be used in a game.</Text>
-              ) : null}
-              <GreenSmallButtonBlock
-                onClick={addTestCase}
-              >
-                Add
-              </GreenSmallButtonBlock>
-            </>
+            <TestCaseEditor
+              newProblem={newProblem}
+              setNewProblem={setNewProblem}
+            />
           ) : null}
 
         {loading ? <Loading /> : null}
